Added a lot of things... should have documented earlier

+++ src/neat/server/server.js
@@ -4,16 +4,20 @@ const fs = require('fs')
 const http = require('http')
 const path = require('path')
 const rimraf = require('rimraf')
+const settings = require('../settings')
+
 const port = process.env.PORT || 3000
 
 let app = express()
 let server = http.Server(app)
 
-app.use(bodyParser.urlencoded({extended: true}))
-app.use('/libs', express.static(path.join(path.resolve(__dirname, '../../'), 'libs')))
-app.use('/dist', express.static(path.join(path.resolve(__dirname, '../../'), 'dist')))
+app.use(bodyParser.urlencoded({extended: true, limit: '50mb'}))
+app.use('/libs', express.static(path.join(__dirname, '../../../libs')))
+app.use('/dist', express.static(path.join(__dirname, '../../../dist')))
 
-rimraf.sync('../training-data')
+if (settings.shouldRunClean) {
+  rimraf.sync('../training-data')
+}
 
 app.get('/', (req, res, next) => {
   res.sendFile(path.join(path.resolve(__dirname, '..'), 'index.html'))
@@ -22,13 +26,20 @@ app.get('/', (req, res, next) => {
 app.post('/store', (req, res, next) => {
   let generation = req.body.generation
   let data = req.body.data
+  let average = req.body.averageFitness
+  let roundHighestFitness = req.body.roundHighestFitness
 
   if (!fs.existsSync('../training-data')) {
     fs.mkdirSync('../training-data')
   }
+  if (!fs.existsSync('../training-data/statistics.csv')) {
+    fs.writeFileSync('../training-data/statistics.csv', '')
+  }
 
-  fs.writeFile(`../training-data/generation-${generation}.json`, data, (err) => {
-    res.json({stored: !err})
+  fs.appendFile('../training-data/statistics.csv', `${generation},${average},${roundHighestFitness}\n`, () => {
+    fs.writeFile(`../training-data/generation-${generation}.json`, data, (err) => {
+      res.json({stored: !err})
+    })
   })
 })
